Refactor to avoid simpleini errors
