+++ libs/config/CMakeLists.txt
@@ -2,11 +2,6 @@
 list(APPEND config_EXTRA_SRCS		"${MRPT_LIBS_ROOT}/config/*.cpp" "${MRPT_LIBS_ROOT}/config/*.h")
 list(APPEND config_EXTRA_SRCS_NAME 	"config" "config")
 
-if (NOT CMAKE_MRPT_HAS_SIMPLEINI_SYSTEM)
-	list(APPEND config_EXTRA_SRCS       "${MRPT_LIBS_ROOT}/config/src/simpleini/*.h")
-	list(APPEND config_EXTRA_SRCS_NAME  "config/simpleini")
-endif()
-
 define_mrpt_lib(
 	config 		# Lib name
 	# Dependencies:
@@ -17,4 +12,7 @@ define_mrpt_lib(
 
 if(BUILD_mrpt-config)
 	target_include_directories(config PRIVATE "${SIMPLEINI_INCLUDE_DIRS}")
+	if (SIMPLEINI_LIBRARY)
+	  target_link_libraries(config PRIVATE ${SIMPLEINI_LIBRARY})
+	endif()
 endif()
