+++ libs/config/src/CConfigFile.cpp
@@ -19,23 +19,41 @@ using binary_function = std::function<RET(T1, T2)>;
 }
 #endif
 
-// SimpleIni: Use Debian icu package instead of copyrighted ConvertUTF.h
-#define SI_CONVERT_ICU 1
-
 #include <SimpleIni.h>
 #include <mrpt/config/CConfigFile.h>
+#include <mrpt/config/config_parser.h>
 #include <mrpt/system/os.h>
-#include "MRPT_SimpleIni.h"
+#include <fstream>
+#include <iostream>
 
 using namespace mrpt;
 using namespace mrpt::config;
-using namespace mrpt::config::simpleini;
 using namespace std;
 
 struct CConfigFile::Impl
 {
-	std::shared_ptr<MRPT_CSimpleIni> ini = std::make_shared<MRPT_CSimpleIni>();
+	std::shared_ptr<CSimpleIniA> ini = std::make_shared<CSimpleIniA>();
 };
+
+// copied from mrpt-io to avoid lib dependency:
+static std::string local_file_get_contents(const std::string& fileName)
+{
+	// Credits: https://stackoverflow.com/a/2602258/1631514
+	// Note: Add "binary" to make sure the "tellg" file size matches the actual
+	// number of read bytes afterwards:
+	std::ifstream t(fileName, ios::binary);
+	if (!t.is_open())
+		THROW_EXCEPTION_FMT(
+			"file_get_contents(): Error opening for read file `%s`",
+			fileName.c_str());
+
+	t.seekg(0, std::ios::end);
+	std::size_t size = t.tellg();
+	std::string buffer(size, ' ');
+	t.seekg(0);
+	t.read(&buffer[0], size);
+	return buffer;
+}
 
 /*---------------------------------------------------------------
 					Constructor
@@ -47,7 +65,10 @@ CConfigFile::CConfigFile(const std::string& fileName)
 
 	m_file = fileName;
 	m_modified = false;
-	m_impl->ini->LoadFile(fileName.c_str());
+
+	const auto sIn = local_file_get_contents(fileName);
+	const auto sOut = mrpt::config::config_parser(sIn);
+	m_impl->ini->LoadData(sOut);
 
 	MRPT_END
 }
@@ -160,10 +181,10 @@ std::string CConfigFile::readString(
  ---------------------------------------------------------------*/
 void CConfigFile::getAllSections(std::vector<std::string>& sections) const
 {
-	MRPT_CSimpleIni::TNamesDepend names;
+	CSimpleIniA::TNamesDepend names;
 	m_impl->ini->GetAllSections(names);
 
-	MRPT_CSimpleIni::TNamesDepend::iterator n;
+	CSimpleIniA::TNamesDepend::iterator n;
 	std::vector<std::string>::iterator s;
 	sections.resize(names.size());
 	for (n = names.begin(), s = sections.begin(); n != names.end(); ++n, ++s)
@@ -176,10 +197,10 @@ void CConfigFile::getAllSections(std::vector<std::string>& sections) const
 void CConfigFile::getAllKeys(
 	const string& section, std::vector<std::string>& keys) const
 {
-	MRPT_CSimpleIni::TNamesDepend names;
+	CSimpleIniA::TNamesDepend names;
 	m_impl->ini->GetAllKeys(section.c_str(), names);
 
-	MRPT_CSimpleIni::TNamesDepend::iterator n;
+	CSimpleIniA::TNamesDepend::iterator n;
 	std::vector<std::string>::iterator s;
 	keys.resize(names.size());
 	for (n = names.begin(), s = keys.begin(); n != names.end(); ++n, ++s)
